refactor(cli): flatten search command module
